+++ LinkEngine/GameTypes/Dungeon/Gameplay/Dungeon.cs
@@ -2,17 +2,25 @@
 
 namespace LinkEngine.Dungeon
 {
-    public class Dungeon
+    public class Dungeon : WorldGen.World
     {
         public List<DungeonItem> ItemsInDungeon { get; set; }
-        
+
+        List<string> PotionImages;
+        List<string> HelmetImages;
+        List<string> TorsoImages;
+        List<string> LegImages;
+        List<string> SwordImages;
+        List<string> ShieldImages;
+
         public Dungeon()
         {
             ItemsInDungeon = new List<DungeonItem>();
-
         }
 
-        // Dungeon Item Creator
+        /// <summary>
+        /// Randomizes the images used for the Icons
+        /// </summary>
         public void RandomizeItems ()
         {
             ItemsInDungeon = new List<DungeonItem>();