+++ LinkEngine/WorldGen/Biome.cs
@@ -6,18 +6,18 @@ namespace LinkEngine.WorldGen
     {
         public int ID { get; set; }
         public string Name { get; set; }
-        public List<Tile> AvailibleTiles { get; set; }
+        public List<Tile> availableTiles { get; set; }
         public List<Biome> RelatedBiomes { get; set; }
 
         public Biome(int id, string name, Tile[] tileAry, Biome[] biomeAry)
         {
             ID = id;
             Name = name;
-            AvailibleTiles = new List<Tile>();
+            availableTiles = new List<Tile>();
 
             for(int i = 0; i < tileAry.Length; i++)
             {
-                AvailibleTiles.Add(tileAry[i]);
+                availableTiles.Add(tileAry[i]);
             }
             if (biomeAry != null)
             {